+++ tigtag3d/Cargo.toml
@@ -10,8 +10,8 @@ bevy.workspace = true
 once_cell.workspace = true
 rand.workspace = true
 chrono.workspace = true
-regex.workspace = true
 
+tigtag3d_inside.workspace = true
 tigtag_inside.workspace = true
 public.workspace = true
