--- tigtag3d/src/tigtag3d/tigtag.rs
@@ -1,6 +1,0 @@
-// use super::*;
-
-//tigtagの名前の輸入
-pub use super::super::tigtag::prelude::*;
-
-//End of code.
