+++ tigtag3d/wasm/tigtag3d.js
@@ -2065,68 +2065,68 @@ function __wbg_get_imports() {
         const ret = wasm.memory;
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_closure_wrapper7416 = function(arg0, arg1, arg2) {
-        const ret = makeMutClosure(arg0, arg1, 5230, __wbg_adapter_34);
+    imports.wbg.__wbindgen_closure_wrapper8634 = function(arg0, arg1, arg2) {
+        const ret = makeMutClosure(arg0, arg1, 5758, __wbg_adapter_34);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_closure_wrapper7417 = function(arg0, arg1, arg2) {
-        const ret = makeMutClosure(arg0, arg1, 5230, __wbg_adapter_34);
+    imports.wbg.__wbindgen_closure_wrapper8635 = function(arg0, arg1, arg2) {
+        const ret = makeMutClosure(arg0, arg1, 5758, __wbg_adapter_34);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_closure_wrapper7418 = function(arg0, arg1, arg2) {
-        const ret = makeMutClosure(arg0, arg1, 5230, __wbg_adapter_34);
+    imports.wbg.__wbindgen_closure_wrapper8636 = function(arg0, arg1, arg2) {
+        const ret = makeMutClosure(arg0, arg1, 5758, __wbg_adapter_34);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_closure_wrapper7419 = function(arg0, arg1, arg2) {
-        const ret = makeMutClosure(arg0, arg1, 5230, __wbg_adapter_41);
+    imports.wbg.__wbindgen_closure_wrapper8637 = function(arg0, arg1, arg2) {
+        const ret = makeMutClosure(arg0, arg1, 5758, __wbg_adapter_41);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_closure_wrapper7420 = function(arg0, arg1, arg2) {
-        const ret = makeMutClosure(arg0, arg1, 5230, __wbg_adapter_34);
+    imports.wbg.__wbindgen_closure_wrapper8638 = function(arg0, arg1, arg2) {
+        const ret = makeMutClosure(arg0, arg1, 5758, __wbg_adapter_34);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_closure_wrapper7421 = function(arg0, arg1, arg2) {
-        const ret = makeMutClosure(arg0, arg1, 5230, __wbg_adapter_46);
+    imports.wbg.__wbindgen_closure_wrapper8639 = function(arg0, arg1, arg2) {
+        const ret = makeMutClosure(arg0, arg1, 5758, __wbg_adapter_46);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_closure_wrapper7422 = function(arg0, arg1, arg2) {
-        const ret = makeMutClosure(arg0, arg1, 5230, __wbg_adapter_34);
+    imports.wbg.__wbindgen_closure_wrapper8640 = function(arg0, arg1, arg2) {
+        const ret = makeMutClosure(arg0, arg1, 5758, __wbg_adapter_34);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_closure_wrapper8169 = function(arg0, arg1, arg2) {
-        const ret = makeMutClosure(arg0, arg1, 5668, __wbg_adapter_51);
+    imports.wbg.__wbindgen_closure_wrapper9387 = function(arg0, arg1, arg2) {
+        const ret = makeMutClosure(arg0, arg1, 6196, __wbg_adapter_51);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_closure_wrapper8170 = function(arg0, arg1, arg2) {
-        const ret = makeMutClosure(arg0, arg1, 5668, __wbg_adapter_54);
+    imports.wbg.__wbindgen_closure_wrapper9388 = function(arg0, arg1, arg2) {
+        const ret = makeMutClosure(arg0, arg1, 6196, __wbg_adapter_54);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_closure_wrapper8171 = function(arg0, arg1, arg2) {
-        const ret = makeMutClosure(arg0, arg1, 5668, __wbg_adapter_54);
+    imports.wbg.__wbindgen_closure_wrapper9389 = function(arg0, arg1, arg2) {
+        const ret = makeMutClosure(arg0, arg1, 6196, __wbg_adapter_54);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_closure_wrapper8172 = function(arg0, arg1, arg2) {
-        const ret = makeMutClosure(arg0, arg1, 5668, __wbg_adapter_54);
+    imports.wbg.__wbindgen_closure_wrapper9390 = function(arg0, arg1, arg2) {
+        const ret = makeMutClosure(arg0, arg1, 6196, __wbg_adapter_54);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_closure_wrapper8173 = function(arg0, arg1, arg2) {
-        const ret = makeMutClosure(arg0, arg1, 5668, __wbg_adapter_54);
+    imports.wbg.__wbindgen_closure_wrapper9391 = function(arg0, arg1, arg2) {
+        const ret = makeMutClosure(arg0, arg1, 6196, __wbg_adapter_54);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_closure_wrapper8174 = function(arg0, arg1, arg2) {
-        const ret = makeMutClosure(arg0, arg1, 5668, __wbg_adapter_54);
+    imports.wbg.__wbindgen_closure_wrapper9392 = function(arg0, arg1, arg2) {
+        const ret = makeMutClosure(arg0, arg1, 6196, __wbg_adapter_54);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_closure_wrapper8175 = function(arg0, arg1, arg2) {
-        const ret = makeMutClosure(arg0, arg1, 5668, __wbg_adapter_54);
+    imports.wbg.__wbindgen_closure_wrapper9393 = function(arg0, arg1, arg2) {
+        const ret = makeMutClosure(arg0, arg1, 6196, __wbg_adapter_54);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_closure_wrapper41090 = function(arg0, arg1, arg2) {
-        const ret = makeMutClosure(arg0, arg1, 30660, __wbg_adapter_67);
+    imports.wbg.__wbindgen_closure_wrapper42308 = function(arg0, arg1, arg2) {
+        const ret = makeMutClosure(arg0, arg1, 31188, __wbg_adapter_67);
         return addHeapObject(ret);
     };
-    imports.wbg.__wbindgen_closure_wrapper81154 = function(arg0, arg1, arg2) {
-        const ret = makeMutClosure(arg0, arg1, 59263, __wbg_adapter_70);
+    imports.wbg.__wbindgen_closure_wrapper82372 = function(arg0, arg1, arg2) {
+        const ret = makeMutClosure(arg0, arg1, 59791, __wbg_adapter_70);
         return addHeapObject(ret);
     };
